Refactor socketcan interface into own package
