+++ can/interfaces/socketcan/__init__.py
@@ -1,0 +1,3 @@
+from can.interfaces.socketcan.socketcan_ctypes import SocketcanCtypes_Bus
+from can.interfaces.socketcan.socketcan_native import SocketcanNative_Bus
+from can.interfaces.socketcan import socketcan_constants as constants
